Move ies implementation to analysis/ies and integrate header files
