+++ libres/lib/analysis/ies/tests/ies_enkf_config.cpp
@@ -1,5 +1,5 @@
 
-#include <ies_enkf_data.hpp>
+#include <ert/analysis/ies/ies_enkf_data.hpp>
 
 void test_create() {
     ies_enkf_config_type *config = ies_enkf_config_alloc();
